tipe data migration

+++ database/migrations/2025_08_06_053005_create_biaya_admins_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->foreignId('id_rate')->constrained('rate');
             $table->foreignId('id_tenor')->constrained('tenor');
-            $table->integer('biaya_admin');
-            $table->integer('min_pinjaman');
-            $table->integer('max_pinjaman');
+            $table->decimal('biaya_admin');
+            $table->decimal('min_pinjaman');
+            $table->decimal('max_pinjaman');
 
 
             $table->timestamps();
